requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: osx-64
_r-mutex=1.0.1=anacondar_1
blas=2.116=openblas
blas-devel=3.9.0=16_osx64_openblas
blosc=1.21.0=h8346a28_1
brotli=1.0.9=hca72f7f_7
brotli-bin=1.0.9=hca72f7f_7
brunsli=0.1=h23ab428_0
bwidget=1.9.14=h694c41f_1
bzip2=1.0.8=h1de35cc_0
c-ares=1.18.1=h0d85af4_0
ca-certificates=2022.10.11=hecd8cb5_0
cairo=1.16.0=he43a7df_1008
cctools_osx-64=973.0.1=h3e07e27_2
certifi=2022.9.24=py39hecd8cb5_0
cfitsio=3.470=hbd21bf8_7
charls=2.2.0=h23ab428_0
clang=12.0.1=h694c41f_4
clang-12=12.0.1=default_he082bbe_4
clang_osx-64=12.0.1=h633439f_9
clangxx=12.0.1=default_he082bbe_4
clangxx_osx-64=12.0.1=hdb584c0_9
cloudpickle=2.0.0=pyhd3eb1b0_0
compiler-rt=12.0.1=he01351e_0
compiler-rt_osx-64=12.0.1=hd3f61c9_0
curl=7.85.0=hca72f7f_0
cycler=0.11.0=pyhd3eb1b0_0
cytoolz=0.12.0=py39hca72f7f_0
dask-core=2022.7.0=py39hecd8cb5_0
eigen=3.4.0=h940c156_0
ffmpeg=4.3.2=h4dad6da_0
fftw=3.3.9=h9ed2024_1
font-ttf-dejavu-sans-mono=2.37=hab24e00_0
font-ttf-inconsolata=3.000=h77eed37_0
font-ttf-source-code-pro=2.038=h77eed37_0
font-ttf-ubuntu=0.83=hab24e00_0
fontconfig=2.13.94=h10f422b_0
fonts-conda-ecosystem=1=0
fonts-conda-forge=1=0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=hd8bbffd_0
fribidi=1.0.10=hbcb3906_0
fsspec=2022.10.0=py39hecd8cb5_0
gettext=0.21.1=h8a4c099_0
gfortran_impl_osx-64=9.3.0=h86db474_7
gfortran_osx-64=9.3.0=h268312c_8
giflib=5.2.1=hbcb3906_2
glib=2.68.4=he49afe7_0
glib-tools=2.68.4=he49afe7_0
gmp=6.2.1=h2e338ed_0
gnutls=3.6.13=h756fd2b_1
graphite2=1.3.14=he9d5cce_1
gsl=2.7=h93259b0_0
gst-plugins-base=1.14.0=h4180768_2
gstreamer=1.14.0=h0fc69c2_2
harfbuzz=2.9.1=h159f659_1
hdf5=1.10.6=h10fe05b_1
icu=68.2=he49afe7_0
imagecodecs=2021.8.26=py39h0f85e6e_1
imageio=2.19.3=py39hecd8cb5_0
isl=0.22.1=hb1e8313_2
jasper=1.900.1=h636a363_1006
joblib=1.1.1=py39hecd8cb5_0
jpeg=9e=hac89ed1_2
jxrlib=1.1=haf1e3a3_2
kiwisolver=1.4.2=py39he9d5cce_0
krb5=1.19.3=hb49756b_0
lame=3.100=hb7f2c08_1003
lcms2=2.12=hf1fd2bf_0
ld64_osx-64=609=hc6b1d1e_10
ldid=2.1.2=h6a69015_3
lerc=3.0=he49afe7_0
libaec=1.0.4=hb1e8313_1
libblas=3.9.0=16_osx64_openblas
libbrotlicommon=1.0.9=hca72f7f_7
libbrotlidec=1.0.9=hca72f7f_7
libbrotlienc=1.0.9=hca72f7f_7
libcblas=3.9.0=16_osx64_openblas
libclang=12.0.0=default_hbc2896b_2
libclang-cpp12=12.0.1=default_he082bbe_4
libcurl=7.85.0=h6dfd666_0
libcxx=14.0.6=h9765a3e_0
libdeflate=1.8=h9ed2024_5
libedit=3.1.20191231=h0678c8f_2
libev=4.33=haf1e3a3_1
libffi=3.3=hb1e8313_2
libgfortran=5.0.0=9_5_0_h97931a8_26
libgfortran5=11.3.0=h082f757_26
libglib=2.68.4=hd556434_0
libiconv=1.17=hac89ed1_0
liblapack=3.9.0=16_osx64_openblas
liblapacke=3.9.0=16_osx64_openblas
libllvm12=12.0.1=hd011deb_2
libnghttp2=1.46.0=ha29bfda_0
libopenblas=0.3.21=openmp_h429af6e_3
libopencv=4.5.3=py39h852ad08_1
libpng=1.6.37=ha441bb4_0
libpq=12.9=h1c9f633_3
libprotobuf=3.16.0=hcf210ce_0
libssh2=1.10.0=h52ee1ee_2
libtiff=4.4.0=h2ef1027_1
libwebp=1.2.4=hfa4350a_0
libwebp-base=1.2.4=h775f41a_0
libxml2=2.9.12=h93ec3fd_0
libzopfli=1.0.3=hb1e8313_0
llvm-openmp=14.0.4=ha654fa7_0
llvm-tools=12.0.1=hd011deb_2
locket=1.0.0=py39hecd8cb5_0
lz4-c=1.9.3=he49afe7_1
make=4.3=h22f3db7_1
matplotlib=3.5.3=py39hecd8cb5_0
matplotlib-base=3.5.3=py39hfb0c5b7_0
mpc=1.2.1=hbb51d92_0
mpfr=4.1.0=h0f52abe_1
munkres=1.1.4=py_0
ncurses=6.3=hca72f7f_3
nettle=3.6=hedd7734_0
networkx=2.8.4=py39hecd8cb5_0
nspr=4.33=he9d5cce_0
nss=3.74=h47edf6a_0
numpy=1.23.4=py39hdfa1d0c_1
openblas=0.3.21=openmp_hbefa662_3
opencv=4.5.3=py39h6e9494a_1
openh264=2.1.1=hfd3ada9_0
openjpeg=2.3.1=h6cbf5cd_3
openssl=1.1.1s=hca72f7f_0
packaging=21.3=pyhd3eb1b0_0
pango=1.48.10=ha05cd14_1
partd=1.2.0=pyhd3eb1b0_1
pcre=8.45=he49afe7_0
pcre2=10.37=ha16e1b2_0
pillow=9.2.0=py39hde71d04_1
pip=22.2.2=py39hecd8cb5_0
pixman=0.40.0=hbcb3906_0
py-opencv=4.5.3=py39h71a6800_1
pyparsing=3.0.9=py39hecd8cb5_0
python=3.9.13=hdfd78df_2
python-dateutil=2.8.2=pyhd3eb1b0_0
python_abi=3.9=2_cp39
pywavelets=1.3.0=py39hca72f7f_0
pyyaml=6.0=py39hca72f7f_1
r-base=4.1.1=h65845f3_1
r-liblinear=2.10_12=r41h49197e3_1
readline=8.2=hca72f7f_0
scikit-image=0.19.2=py39hae1ba45_0
scikit-learn=1.0.2=py39hae1ba45_1
scipy=1.9.3=py39h1d874dc_0
setuptools=65.5.0=py39hecd8cb5_0
sigtool=0.1.3=h57ddcff_0
six=1.16.0=pyhd3eb1b0_1
snappy=1.1.9=he9d5cce_0
sqlite=3.39.3=h707629a_0
tapi=1100.0.11=h9ce4665_0
threadpoolctl=2.2.0=pyh0d69192_0
tifffile=2021.7.2=pyhd3eb1b0_2
tk=8.6.12=h5d9f67b_0
tktable=2.10=h49f0cf7_3
toolz=0.12.0=py39hecd8cb5_0
tornado=6.2=py39hca72f7f_0
tzdata=2022f=h04d1e81_0
wheel=0.37.1=pyhd3eb1b0_0
x264=1!161.3030=h0d85af4_1
xz=5.2.6=hca72f7f_0
yaml=0.2.5=haf1e3a3_0
zfp=0.5.5=he9d5cce_6
zlib=1.2.13=h4dc903c_0
zstd=1.5.2=hcb37349_0
